+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.3] - 2024-04-01
+
+### Fixed
+- Update git command to get commit time to support annotated tags
+
 ## [0.4.2] - 2024-03-26
 
 ### Changed